refactor: add wait calls
